because of Laravel 9 updates
Also removed more CSS that was no longer in use

+++ app/Http/Controllers/User/DestroyController.php
@@ -20,7 +20,9 @@ class DestroyController extends Controller
     {
         $user = $request->user();
 
-        Storage::delete($user->avatar);
+        if ($user->avatar) {
+            Storage::delete($user->avatar);
+        }
 
         $subject = 'Account Deletion.';
         $message = 'Your account has been deleted.';
